Restructure into new plants feature
