+++ client/src/features/calendar/routes/Home.tsx
@@ -1,15 +1,15 @@
-import Calendar from "../components/Calendar";
+import { Loader } from "@/components/Loader";
+import { useDesktopWidth } from "@/hooks/useDesktopWidth";
+import { ErrorPage } from "@/routes/ErrorPage";
 import { useQuery } from "@tanstack/react-query";
 import { getAllPlants } from "../api";
-import { Info } from "../components/Info";
-import { useShowFormStore } from "../stores/showFormStore";
+import Calendar from "../components/Calendar";
 import { Filter } from "../components/Filter";
 import { FilterForm } from "../components/Filter/FilterForm";
+import { Info } from "../components/Info";
 import { useFilterStore } from "../stores/filterStore";
-import { Loader } from "@/components/Loader";
-import { ErrorPage } from "@/routes/ErrorPage";
-import { Plants } from "./Plants";
-import { useDesktopWidth } from "@/hooks/useDesktopWidth";
+import { useShowFormStore } from "../stores/showFormStore";
+import { PlantsRoute } from "../../plants/routes/PlantsRoute";
 
 export const Home = () => {
   const { data, isLoading, isError } = useQuery({
@@ -19,6 +19,7 @@ export const Home = () => {
   const filterSelections = useFilterStore((state) =>
     state.plants.map((plant) => plant.id)
   );
+
   const showFilterForm = useShowFormStore((state) => state.filterPlant);
 
   const [width] = useDesktopWidth();
@@ -38,6 +39,7 @@ export const Home = () => {
 
       <div className="relative flex h-full w-full max-w-md flex-col gap-4 ">
         <Filter />
+        {/* todo move this inside Info */}
         {showFilterForm ? (
           <FilterForm plants={data} />
         ) : (
@@ -51,7 +53,7 @@ export const Home = () => {
         )}
       </div>
       {/* Fix using media query */}
-      {width && <Plants />}
+      {width && <PlantsRoute />}
     </section>
   );
 };
