+++ client/src/main.tsx
@@ -16,7 +16,7 @@ import { User } from "./features/user/routes/User";
 import { checkSession } from "./utils/checkSession";
 import { PasswordForm } from "@/features/auth/components/Forms/PasswordForm";
 import { NewPasswordForm } from "@/features/auth/components/Forms/NewPasswordForm";
-import { Plants } from "@/features/calendar/routes/Plants";
+import { PlantsRoute } from "@/features/plants/routes/PlantsRoute";
 import { Layout } from "./routes/Layout";
 
 const queryClient = new QueryClient();
@@ -40,7 +40,7 @@ const router = createBrowserRouter([
     path: "/plants",
     element: (
       <Layout>
-        <Plants />
+        <PlantsRoute />
       </Layout>
     ),
     loader: checkSession,
